Version 1.41
Change the behavior of :IndentRangeConsistencyCop to consider the buffer settings to turn around the verdict of "inconsistent indent" (but still not report inconsistent buffer settings alone). Otherwise, together with the IndentConsistencyCopAutoCmds triggers, it can happen that on opening (i.e. :IndentConsistencyCop), the file is judged okay (considering the buffer settings), but on writing the buffer (:IndentRangeConsistencyCop), a potential inconsistency due to too small indent is reported. Thanks to Marcelo Montú for reporting this issue.