Version 1.42
When a perfect or authoritative rating didn't pass the majority rule, try to turn around the verdict by checking consistency with buffer settings, as is done for small indents only. For example, this avoids a wrong verdict of inconsistent spc8 when there are more spc8 than spc4. Thanks to Marcelo Montú for reporting this issue.